+++ tests/tests_main.c
@@ -35,6 +35,7 @@ int main(int argc, char** argv)
     // registerTest("Input Window Test", inputWindowTest);
     // registerTest("System Cursor Test", systemCursorTest);
     // registerTest("Attach Window Test", attachWindowTest);
+    registerTest("Character Event Test", charEventTest);
 #endif // PAL_HAS_VIDEO
 
 #if PAL_HAS_OPENGL
@@ -50,7 +51,7 @@ int main(int argc, char** argv)
 #endif // PAL_HAS_OPENGL
 
 #if PAL_HAS_OPENGL && PAL_HAS_VIDEO && PAL_HAS_THREAD
-    registerTest("Multi Thread OpenGL Test", multiThreadOpenGlTest);
+    //registerTest("Multi Thread OpenGL Test", multiThreadOpenGlTest);
 #endif //
 
     runTests();